requirements.txt
flask
astro-datalab
google-genai
beautifulsoup4
requests
numpy
scikit-learn
sentence-transformers
